+++ CHANGELOG.md
@@ -38,5 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+- [#108](https://github.com/babylonlabs-io/babylon-sdk/pull/108) chore: Upgrade wasmd to 0.55
 - [#116](https://github.com/babylonlabs-io/babylon-sdk/pull/116) add docs for
   the x/babylon module